requirements.txt
Jinja2==3.1.4
MarkupSafe==2.1.5
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
python-slugify==8.0.4
Markdown==3.6
python-dotenv==1.0.1
anthropic>=0.40.0
